Support running app as Docker image
This change will allow for both local development (make dev) as well as
running the application as a container with Docker (make start). The
docker solution uses uWSIG as the server. We do not use Flask's server
in production because it's built to simply be a development server, it's
not built for scaling and high traffic.